requirements.txt
Gunicorn
Flask
Flask-SQLAlchemy
PyMySQL
Flask-CORS
python-dotenv
flask-cors
Flask-Mail
cryptography
pandas